requirements.txt
streamlit
torch==2.3.0
torchvision==0.18.0
numpy
pandas
pillow